Chore: Project view page refactored
