+++ src/components/Project/NotesFeed.jsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import Note from './Note';
 
-import styles from '../../styles/noteFeed.module.scss';
-
 export default ({ notes }) => {
   console.log(notes);
   if (notes && notes.length) {
     console.log(notes);
     return (
-      <div className={styles['notes-container']}>
+      <div>
         {notes.map(note => {
           return <Note note={note} />;
         })}
