--- src/styles/dashboard.module.scss
@@ -1,14 +1,0 @@
-@import 'variables.scss';
-
-.dashboard-parent-container {
-    display: flex;
-}
-
-.dashboard-sidebar {
-    position: fixed;
-}
-
-.dashboard-project-list {
-    margin: 2rem 2rem 2rem 15rem ;
-    width: 100%;
-}
