--- src/styles/editor.module.scss
@@ -1,132 +1,0 @@
-@import 'variables.scss';
-
-h2 {
-  font-size: 1.4rem;
-  font-family: $fira;
-}
-.main-container {
-  margin-top: 1rem;
-  
-  .editor-container {
-    margin-top: 25px;
-    display: flex;
-    .avatar-container {
-      padding: 0.5rem;
-      img {
-        width: 50px;
-        height: 50px;
-        border-radius: 50%;
-      }
-    }
-    .form-container {
-      width: 35rem;
-      background: rgba(33, 34, 47, 0.02);
-      border: solid 0.05rem rgba(33, 34, 47, 0.2);
-      padding: 2% 3%;
-      border-radius: 1rem;
-      input {
-        background: inherit;
-        border: none;
-      }
-      input:focus {
-        outline: none;
-      }
-      textarea {
-        background: inherit;
-        padding: 0;
-        border: none;
-        min-width: 400px;
-        max-width: 100%;
-        min-height: 20vh;
-        resize: none;
-        &:focus {
-          outline: none;
-        }
-      }
-    }
-    .form-header {
-      display: flex;
-      justify-content: space-between;
-      margin-bottom: 2vh;
-    }
-    .body-input {
-      height: 15vh;
-    }
-    .text-footer {
-      display: flex;
-      flex-flow: row;
-      justify-content: space-between;
-      align-items: center;
-      width: 100%;
-      p {
-        font-size: 0.6rem;
-      }
-      img {
-        width: 25px;
-        border-radius: 50%;
-      }
-      .expanded {
-        .attendees-avatars {
-          display: flex;
-          margin-top: 0.5vh;
-          flex-flow: wrap;
-        }
-        .mini-avatar-container {
-          display: flex;
-          align-items: center;
-          width: 40%;
-          p {
-            overflow: hidden;
-            text-overflow: ellipsis;
-            white-space: nowrap;
-            margin-left: 0.5rem;
-          }
-          button {
-            border: none;
-            background: none;
-          }
-        }
-      }
-      .collapsed {
-        .attendees-avatars {
-          display: flex;
-          margin-top: 0.5vh;
-          img {
-            margin-right: -6px;
-            box-shadow: -2px 2px rgba(33, 34, 47, 0.1);
-          }
-          p {
-            display: none;
-          }
-          button {
-            display: none;
-          }
-        }
-      }
-      .button-container {
-        display: flex;
-        justify-content: space-between;
-        height: 4.5vh;
-        max-height: 40px;
-        .attach-btn {
-          width: 5rem;
-          color: $accent;
-          border: none;
-          background: inherit;
-          svg {
-            path {
-              color: $accent;
-            }
-          }
-        }
-        .save-btn {
-          background: $accent;
-          border-radius: 8%;
-          color: white;
-          width: 5rem;
-          border: none;
-        }
-      }
-    }
-  }
-}
