--- src/styles/layout.module.scss
@@ -1,10 +1,0 @@
-@import 'variables.scss';
-
-.main-container {
-  width: 100%;
-
-  .content-container {
-    margin-left: 10rem;
-    padding: 20px 40px;
-  }
-}
