--- src/styles/notes.module.scss
@@ -1,111 +1,0 @@
-@import 'variables.scss';
-
-section {
-  display: flex;
-  width: 42rem;
-  margin-bottom: 10px;
-  padding-left: 0.5rem;
-}
-
-.avatar-container {
-  width: 75px;
-}
-
-.avatar {
-  border-radius: 50%;
-  width: 50px;
-  height: 50px;
-  margin: 10px 0;
-}
-
-.content-container {
-  display: flex;
-  flex-flow: column;
-}
-
-.expanded {
-  display: flex;
-  flex-wrap: wrap;
-  .mini-avatar-container {
-    width: 40%;
-    display: flex;
-    flex-flow: row;
-    p {
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-    }
-  }
-}
-
-.note-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-  padding-bottom: 20px;
-  padding-top: 10px;
-  .note-header {
-    display: flex;
-    justify-content: space-between;
-    font-size: 0.75rem;
-    align-items: center;
-  }
-}
-
-section:not(:last-child) {
-  .note-container {
-    border-bottom: 2px solid rgba(0, 0, 0, 0.1);
-  }
-}
-
-.note-body {
-  margin-bottom: 0.7rem;
-  font-size: 0.7rem;
-  color: $copy;
-  line-height: 1.2rem;
-}
-
-.note-footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  button {
-    align-items: flex-start;
-    background: none;
-    border: none;
-    color: $copy;
-    font-size: 0.7rem;
-    height: 25px;
-    width: 45px;
-  }
-}
-
-.mini-avatar-container {
-  display: flex;
-  flex-wrap: wrap;
-  img {
-    height: 30px;
-    width: 30px;
-    border-radius: 50%;
-    margin-top: 10px;
-  }
-  p {
-    font-size: 0.6rem;
-    margin: 18px 0 0 5px;
-  }
-}
-
-.collapsed {
-  display: flex;
-  img {
-    margin-right: -6px;
-    box-shadow: -2px 2px $light;
-  }
-  p {
-    display: none;
-  }
-  button {
-    display: none;
-  }
-}
