--- src/styles/projectHeader.module.scss
@@ -1,97 +1,0 @@
-@import 'variables.scss';
-
-* {
-    font-family: $noto;
-    font-size: 1rem;
-    color: $dark;
-}
-
-.project-header-container {
-    display: flex;
-    flex-direction: column;
-    padding: 4vh 0 0;
-
-    .project-back-link  {
-        font-family: $fira;
-        background-color: white;
-        text-decoration: none;
-        color: $accent;
-        font-weight: bold;
-    }
-}
-
-.project-summary-container {
-    display: flex;
-
-    .project-title-container {
-        width: 48rem;
-    
-        .project-title {
-            margin: 10px 0;
-            font-size: 2.5rem;
-        }
-        
-        .project-cohort {
-            margin: 10px 0;
-            border: 1px solid #71727a;
-            border-radius: 9px;
-            width: 100px;
-            text-align: center;
-            font-size: 0.6rem;
-            padding: 3px;
-            color: #71727a;
-            cursor: default;
-        }
-    }
-    
-    .project-status-container {
-        display: flex;
-        justify-content: center;
-        margin-top: 2rem;
-    
-        .project-status {
-            width: 50%;
-            font-size: 1.1rem;
-            font-weight: bold;
-            margin: 0 5%;
-        }
-        
-        .project-fire {
-            width: 40px;
-            height: 55px;
-            box-sizing: border-box;
-            padding-bottom: 5px; 
-        }
-    }
-}
-
-.project-tabs-container {
-    display: flex;
-    margin: 0;
-
-    .project-overview-tab {
-        font-family: $fira;
-        margin: 5px 20px 0 0;
-        cursor: pointer;
-        padding: 0 2px 4px 1px;
-        border-bottom: 2px solid $accent;
-        font-weight: bold;
-    }
-
-    .project-dynamics-tab {
-        font-family: $fira;
-        margin: 5px 20px 0 0;
-        cursor: pointer;
-        padding: 0 2px 4px 1px;
-    }
-    .project-dynamics-tab:hover {
-        border-bottom: 2px solid $accent;
-        padding: 0 2px 2px 1px;
-    }
-}
-
-.project-content-divider {
-    display: flex;
-    opacity: 0.2;
-    margin: 0 10% 0 0; 
-}
