--- src/styles/projectList.module.scss
@@ -1,77 +1,0 @@
-@import 'variables.scss';
-
-th,
-td,
-a {
-  font-family: $noto;
-  font-size: 1rem;
-  color: $dark;
-}
-
-a {
-  text-decoration: none;
-}
-
-table {
-  margin-top: 4rem;
-  width: 100%;
-  border-collapse: collapse;
-}
-
-td {
-  padding: 20px 20px;
-  border-bottom: 2px solid rgba(33, 34, 47, 0.05);
-  text-align: left;
-}
-
-.pl {
-  // padding-right: 20rem;
-}
-
-th {
-  text-transform: uppercase;
-  text-align: left;
-  padding: 8px 20px;
-  font-size: 0.7rem;
-  font-weight: 900;
-  border-bottom: 2px solid rgba(33, 34, 47, 0.16);
-}
-
-.rtd {
-  text-align: right;
-}
-
-.rtc {
-  text-align: center;
-  padding-top: 0;
-  padding-bottom: 0;
-  padding-right: 0;
-  padding-left: 50px;
-
-  .rtci {
-    height: 40px;
-  }
-}
-
-tr:nth-child(even) {
-  background: $light;
-}
-
-.title {
-  color: $accent;
-  font-weight: 900;
-}
-
-// button {
-//   border-radius: 5px;
-//   padding: 5px 20px 6px;
-//   transition: 600ms;
-// }
-
-// button:hover {
-//   transition: 600ms;
-//   transform: scale(1.1);
-//   color: white;
-//   background-color: $accent;
-//   cursor: pointer;
-// }
