--- src/styles/projectView.module.scss
@@ -1,30 +1,0 @@
-@import 'variables.scss';
-
-.parent-project-container {
-  display: flex;
-  margin: 0 auto;
-  width: 1000px;
-}
-
-.project-page-sidebar {
-  margin-right: 5%;
-}
-
-.project-page-contents {
-  width: 100%;
-}
-
-.project-container {
-  display: flex;
-}
-
-.editor-feed-container {
-  display: flex;
-  flex-flow: column;
-  justify-content: space-around;
-  min-width: 30rem;
-}
-
-.team-container {
-  margin-left: 8%;
-}
