--- src/styles/sidebar.module.scss
@@ -1,89 +1,0 @@
-@import 'variables.scss';
-
-.sidebar {
-  display: flex;
-  flex-direction: column;
-  min-width: 10rem;
-  background-color: #f9f9f9;
-  height: 100vh;
-  position: fixed;
-}
-
-.logo {
-  padding: 2rem;
-  padding-left: 3rem;
-  width: 35px;
-}
-
-.orgIcon {
-  background-color: $accent;
-  height: 50px;
-  width: 50px;
-  margin: 2rem;
-}
-
-.links {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: space-between;
-}
-
-.link {
-  font-family: $fira;
-  text-decoration: none;
-  padding-left: 3rem;
-  border-left: 5px solid #f9f9f9;
-  margin-bottom: 20px;
-}
-
-.link:hover {
-  color: $accent;
-}
-
-.current {
-  border-left: 6px solid $accent;
-  font-weight: 900;
-  margin-bottom: 20px;
-}
-
-.topLinks {
-  display: flex;
-  flex-direction: column;
-}
-
-.bottomLinks {
-  display: flex;
-  flex-direction: column;
-  padding-bottom: 10px;
-}
-
-.logoutButton {
-  padding: 0.5rem 1rem;
-  -webkit-transition: background 0.15s ease-in-out;
-  transition: background 0.15s ease-in-out;
-  -webkit-transition: color 0.1s ease-in-out;
-  transition: color 0.1s ease-in-out;
-  background: $accent;
-  border: 1px solid $accent;
-  border-radius: 3px;
-  color: $light;
-  font-family: $noto;
-  font-size: 1.2rem;
-  width: 8rem;
-  margin: 0 auto;
-
-  a {
-    color: $light;
-  }
-}
-
-.logoutButton:hover {
-  background: transparent;
-  color: $accent;
-  cursor: pointer;
-
-  a {
-    color: $accent;
-  }
-}
