--- src/styles/team.module.scss
@@ -1,49 +1,0 @@
-@import 'variables.scss';
-
-.avatar {
-  border-radius: 50%;
-  width: 40px;
-  height: 40px;
-  margin: 10px 0;
-}
-
-.team-names {
-  color: $dark;
-  margin-left: 8px;
-}
-
-.members {
-  display: flex;
-  align-items: center;
-}
-
-.member-container {
-  display: flex;
-  flex-direction: column;
-  padding-bottom: 5px;
-  width: max-content;
-
-  h3 {
-    font-family: $fira;
-    font-style: normal;
-    font-weight: bold;
-    font-size: 12px;
-    line-height: 14px;
-    text-transform: uppercase;
-    color: $copy;
-  }
-}
-
-.member-container:nth-child(3) {
-  padding-bottom: 10px;
-  border-bottom: 2px solid rgba(0, 0, 0, 0.1);
-}
-
-.team-container {
-  margin-left: 5%;
-  margin-top: 1rem;
-  h2 {
-    font-size: 1.4rem;
-    font-family: $fira;
-  }
-}
